Move all move functions to move.hpp, remove move namespace
