+++ src/test/TypesTests.cpp
@@ -1,6 +1,7 @@
 #include <catch2/catch_test_macros.hpp>
 
 #include <types.h>
+#include <move.hpp>
 
 namespace yak {
